searchy/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary  # For PostgreSQL connection
langchain
langgraph
langchain-groq
langchain-community # For loaders, splitters, embeddings
faiss-cpu        # Or faiss-gpu if you have CUDA configured
sentence-transformers
python-dotenv
pydantic-settings # For loading .env into Pydantic models
requests
beautifulsoup4
duckduckgo-search
tiktoken # Often a dependency for text splitters
numpy # Dependency for FAISS/embeddings
pickle-mixin # Sometimes needed for saving FAISS metadata reliably